prepare docs for cog model
